refactor : 서빙시스템 모듈명 변
