+++ rtw-serve/src/main/kotlin/com/javacafe/rtwserve/common/config/AopConfig.kt
@@ -1,0 +1,8 @@
+package com.javacafe.rtwserve.common.config
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.context.annotation.EnableAspectJAutoProxy
+
+@Configuration
+@EnableAspectJAutoProxy
+class AopConfig
