+++ rtw-serve/src/main/kotlin/com/javacafe/rtwserve/config/ElasticsearchConfig.kt
@@ -1,0 +1,10 @@
+package com.javacafe.rtwserve.config
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.data.elasticsearch.repository.config.EnableElasticsearchRepositories
+
+@Configuration
+@EnableElasticsearchRepositories(
+    basePackages = ["com.javacafe.rtwserve.domain.*.repository"]
+)
+class ElasticsearchConfig
