+++ rtw-serve/src/test/kotlin/com/javacafe/rtwserve/RtwServeApplicationTests.kt
@@ -1,0 +1,13 @@
+package com.javacafe.rtwserve
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class RtwServeApplicationTests {
+
+    @Test
+    fun contextLoads() {
+    }
+
+}
