+++ rtw-serve/src/test/resources/http/news-api-test.http
