--- rtw-serving/src/test/kotlin/com/javacafe/rtwserving/RtwServingApplicationTests.kt
@@ -1,13 +1,0 @@
-package com.javacafe.rtwserving
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class RtwServingApplicationTests {
-
-    @Test
-    fun contextLoads() {
-    }
-
-}
