+++ settings.gradle.kts
@@ -1,4 +1,6 @@
 rootProject.name = "realtime-winnmin"
 
 include("rtw-collector")
-include("rtw-serving")
+include("rtw-serve")
+include("rtw-dataclean")
+include("rtw-index")
